+++ previous-versions/files/RELEASES.md
@@ -595,3 +595,13 @@ Based on:
 - [python v0.1.0] previous-versions/files
 ### Releases
 - [PyPI v0.1.0] https://pypi.org/project/codat-files/0.1.0 - previous-versions/files
+
+## 2023-09-13 11:12:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Files.yaml
+- Speakeasy CLI 1.82.5 (2.108.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.1.0] previous-versions/files
+### Releases
+- [PyPI v0.1.0] https://pypi.org/project/codat-files/0.1.0 - previous-versions/files